requirements.txt
astroid==2.2.3
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.2rc1
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
idna==2.8
isort==4.3.10
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
